requirements-test.txt
# Testing dependencies for green-ai-carbon-scheduler
# Install with: pip install -r requirements-test.txt

# Core testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Code quality
pytest-flake8>=1.1.1
pytest-pylint>=0.21.0
black>=23.0.0
isort>=5.12.0

# Coverage reporting
coverage>=7.2.0

# Mock and testing utilities
requests-mock>=1.11.0

# Optional: for better test output
pytest-xdist>=3.3.0  # Parallel test execution
pytest-timeout>=2.1.0  # Timeout handling
